Finalize and document engine implementation
