+++ optd-dsl/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 optd-core = { path = "../optd-core" }
 chumsky = "0.9.3"
 ariadne = "0.5.0"
-ordered-float = "4.6.0"
+ordered-float = "5.0.0"
 futures = "0.3.31"
 tokio.workspace = true
+async-recursion.workspace = true
